+++ src/main/java/ph/devcon/rapidpass/services/ICheckpointService.java
@@ -5,7 +5,6 @@ import ph.devcon.rapidpass.entities.ScannerDevice;
 
 public interface ICheckpointService
 {
-    AccessPass retrieveAccessPassByControlCode(String controlCode);
     AccessPass retrieveAccessPassByPlateNo(String plateNo);
     AccessPass retrieveAccessPassByQrCode(String qrCode);
     ScannerDevice retrieveDeviceByImei(String imei);