+++ src/main/java/ph/devcon/rapidpass/utilities/ControlCodeGenerator.java
@@ -42,4 +42,28 @@ public class ControlCodeGenerator {
         int checkDigit = Damm32.compute(obfuscatedId);
         return CrockfordBase32.encode(obfuscatedId, 7) + CrockfordBase32.encode(checkDigit);
     }
+
+    public static int decode(String qrCodeSecretKeyInBase64, String controlCode) {
+
+        // The secret key in base 64 format.
+        byte[] qrCodeSecretKey = Base64.getDecoder().decode(qrCodeSecretKeyInBase64);
+
+        if (qrCodeSecretKey.length < 10) {
+            /*
+             * The qrCodeSecretKey should be at least 10 characters, as this is required by Skip32.
+             */
+            throw new IllegalArgumentException("Failed to generate QR code because the secret is not configured properly (key should be at least 10 characters, found: " + qrCodeSecretKey.length + " length)");
+        }
+
+        int length = controlCode.length();
+        String base = controlCode.substring(0, length - 1);
+        String actual = controlCode.substring(length - 1, length);
+
+        long obfuscatedId = CrockfordBase32.decode(base);
+
+        int decryptedNumber = Skip32.decrypt((int) obfuscatedId, qrCodeSecretKey);
+
+        return decryptedNumber;
+
+    }
 }