+++ src/test/java/ph/devcon/rapidpass/services/QrPdfServiceTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import ph.devcon.dctx.rapidpass.commons.HmacSha256;
 import ph.devcon.dctx.rapidpass.commons.QrCodeSerializer;
@@ -15,6 +16,8 @@ import ph.devcon.dctx.rapidpass.commons.Signer;
 import ph.devcon.dctx.rapidpass.model.ControlCode;
 import ph.devcon.rapidpass.entities.AccessPass;
 import ph.devcon.rapidpass.repositories.AccessPassRepository;
+import ph.devcon.rapidpass.services.controlcode.ControlCodeService;
+import ph.devcon.rapidpass.services.controlcode.ControlCodeServiceImpl;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
@@ -28,8 +31,7 @@ import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.fail;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.when;
 
 /**
@@ -43,12 +45,15 @@ class QrPdfServiceTest {
 
     QrPdfService instance;
 
-    @Mock
-    AccessPassRepository accessPassRepository;
+    @Mock private AccessPassRepository accessPassRepository;
 
+    @Mock private ControlCodeService controlCodeService;
+
+    @Mock  private QrGeneratorServiceImpl qrGenService;
 
     private String encryptionKey = "***REMOVED***";
     private String signingKey = "***REMOVED***";
+    ;
 
     @BeforeEach
     void setUp() {
@@ -57,8 +62,13 @@ class QrPdfServiceTest {
         final QrCodeSerializer qrCodeSerializer = new QrCodeSerializer(encryptionKeyBytes);
         final Signer signer = HmacSha256.signer(signingKeyBytes);
 
-        final QrGeneratorServiceImpl qrGenService = new QrGeneratorServiceImpl(qrCodeSerializer, signer);
-        instance = new QrPdfService(qrGenService, accessPassRepository);
+        qrGenService = new QrGeneratorServiceImpl(qrCodeSerializer, signer);
+
+        controlCodeService = Mockito.mock(ControlCodeService.class);
+
+        accessPassRepository = Mockito.mock(AccessPassRepository.class);
+
+        instance = new QrPdfService(qrGenService, accessPassRepository, controlCodeService);
 
     }
 
@@ -73,11 +83,13 @@ class QrPdfServiceTest {
 
     @Test
     void generateQrPdf_INDIVIDUAL() throws IOException, WriterException, ParseException, NullPointerException {
+        instance = new QrPdfService(qrGenService, accessPassRepository, controlCodeService);
 
         String controlCode = ControlCode.encode(38);
 
-        AccessPass accessPass = AccessPass.builder().
-                status("APPROVED")
+        AccessPass accessPass = AccessPass.builder()
+                .id(38)
+                .status("APPROVED")
                 .referenceID("09171234567")
                 .passType("INDIVIDUAL")
                 .controlCode(controlCode)
@@ -90,8 +102,12 @@ class QrPdfServiceTest {
                 .validTo(NOW.plusDays(1))
                 .build();
 
-        when(accessPassRepository.findByControlCode(anyString()))
+        when(controlCodeService.findAccessPassByControlCode(anyString()))
                 .thenReturn(accessPass);
+
+        when(controlCodeService.encode(anyInt())).thenReturn(accessPass.getControlCode());
+
+//        when(this.qrGenService.generateQr(any())).thenReturn(new byte[]{ 0, 1, 0, 1, 0 });
 
         final byte[] bytes = ((ByteArrayOutputStream) instance.generateQrPdf(accessPass.getControlCode())).toByteArray();
 
@@ -107,8 +123,9 @@ class QrPdfServiceTest {
 
         String controlCode = ControlCode.encode(38);
 
-        AccessPass accessPass = AccessPass.builder().
-                status("APPROVED")
+        AccessPass accessPass = AccessPass.builder()
+                .id(38)
+                .status("APPROVED")
                 .referenceID("09171234567")
                 .passType("VEHICLE")
                 .controlCode(controlCode)
@@ -122,8 +139,11 @@ class QrPdfServiceTest {
                 .validTo(NOW.plusDays(1))
                 .build();
 
-        when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
+        when(controlCodeService.findAccessPassByControlCode(any()))
                 .thenReturn(accessPass);
+
+        when(controlCodeService.encode(anyInt()))
+                .thenReturn(controlCode);
 
         final byte[] bytes = ((ByteArrayOutputStream) instance.generateQrPdf(accessPass.getControlCode())).toByteArray();
         writeBytesForVisualInspection(bytes);
@@ -153,10 +173,6 @@ class QrPdfServiceTest {
                     .validTo(NOW.plusDays(1))
                     .build();
 
-
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
-
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");
         });
@@ -185,9 +201,6 @@ class QrPdfServiceTest {
                     .build();
 
 
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
-
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");
         });
@@ -213,10 +226,6 @@ class QrPdfServiceTest {
                     .validTo(NOW.plusDays(1))
                     .build();
 
-
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
-
             instance.generateQrPdf(accessPass.getControlCode());
         });
 
@@ -236,10 +245,6 @@ class QrPdfServiceTest {
                     .validFrom(NOW)
                     .validTo(NOW.plusDays(1))
                     .build();
-
-
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
 
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");
@@ -268,10 +273,6 @@ class QrPdfServiceTest {
                     .validTo(NOW.plusDays(1))
                     .build();
 
-
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
-
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");
         });
@@ -298,10 +299,6 @@ class QrPdfServiceTest {
                     .validTo(NOW.plusDays(1))
                     .build();
 
-
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
-
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");
         });
@@ -323,10 +320,6 @@ class QrPdfServiceTest {
                     .validFrom(NOW)
                     .validTo(NOW.plusDays(1))
                     .build();
-
-
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
 
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");
@@ -356,9 +349,6 @@ class QrPdfServiceTest {
                     .build();
 
 
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
-
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");
         });
@@ -381,9 +371,6 @@ class QrPdfServiceTest {
                     .validTo(NOW.plusDays(1))
                     .build();
 
-
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
 
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");
@@ -412,10 +399,6 @@ class QrPdfServiceTest {
                     .validTo(NOW.plusDays(1))
                     .build();
 
-
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
-
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");
         });
@@ -437,11 +420,6 @@ class QrPdfServiceTest {
                     .validFrom(NOW)
                     .validTo(NOW.plusDays(1))
                     .build();
-
-
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
-
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");
         });
@@ -468,10 +446,6 @@ class QrPdfServiceTest {
                     .validTo(NOW.plusDays(1))
                     .build();
 
-
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
-
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");
         });
@@ -494,10 +468,6 @@ class QrPdfServiceTest {
                     .validFrom(NOW)
                     .validTo(NOW.plusDays(1))
                     .build();
-
-
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
 
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");
@@ -525,10 +495,6 @@ class QrPdfServiceTest {
                     .validTo(NOW.plusDays(1))
                     .build();
 
-
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
-
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");
         });
@@ -551,10 +517,6 @@ class QrPdfServiceTest {
                     .validFrom(NOW)
                     .validTo(INVALID_ARGUMENT)
                     .build();
-
-
-            when(accessPassRepository.findByControlCode(eq(accessPass.getControlCode())))
-                    .thenReturn(accessPass);
 
             instance.generateQrPdf(accessPass.getControlCode());
             fail("should throw exception");