+++ CHANGELOG.md
@@ -30,6 +30,7 @@ CHECKPOINT
 
 ## Changed
 - default expiration date is now April 30, 2020
+- Control code generation now performed on-the-fly, rather than queried from the database.
 
 ## [1.1.1] - 2020-04-07
 ### Changed